Refactor input verification code
