+++ cosmwasm/enclaves/shared/contract-engine/src/contract_validation.rs
@@ -8,16 +8,16 @@ use cw_types_generic::BaseEnv;
 use cw_types_v010::types::{CanonicalAddr, Coin, HumanAddr};
 use enclave_cosmos_types::traits::CosmosAminoPubkey;
 use enclave_cosmos_types::types::{
-    is_transfer_ack_error, ContractCode, CosmosPubKey, CosmosSdkMsg, FungibleTokenPacketData,
-    HandleType, IBCLifecycleComplete, IBCLifecycleCompleteOptions, IBCPacketAckMsg,
-    IBCPacketTimeoutMsg, IbcHooksIncomingTransferMsg, IbcHooksOutgoingTransferMemo, Packet,
-    SigInfo, SignDoc, StdSignDoc,
+    ContractCode, CosmosPubKey, CosmosSdkMsg, HandleType, SigInfo, SignDoc, StdSignDoc,
 };
 use enclave_crypto::traits::VerifyingKey;
 use enclave_crypto::{sha_256, AESKey, Hmac, Kdf, HASH_SIZE, KEY_MANAGER};
 use enclave_ffi_types::EnclaveError;
 
-use crate::ibc_denom_utils::{get_denom_prefix, parse_denom_trace, receiver_chain_is_source};
+use crate::input_validation::contract_address_validation::verify_contract_address;
+use crate::input_validation::msg_validation::verify_and_get_sdk_msg;
+use crate::input_validation::send_funds_validations::verify_sent_funds;
+use crate::input_validation::sender_validation::verify_sender;
 use crate::io::create_callback_signature;
 use crate::message::is_ibc_msg;
 use crate::types::SecretMessage;
@@ -637,489 +637,6 @@ fn verify_callback_sig_impl(
     true
 }
 
-/// Get the cosmwasm message that contains the encrypted message
-fn verify_and_get_sdk_msg<'sd>(
-    messages: &'sd [CosmosSdkMsg],
-    msg_sender: &CanonicalAddr,
-    sent_msg: &SecretMessage,
-    handle_type: HandleType,
-) -> Option<&'sd CosmosSdkMsg> {
-    trace!("get_verified_msg: {:?}", messages);
-
-    messages.iter().find(|&m| match m {
-        CosmosSdkMsg::MsgExecuteContract { msg, sender, .. }
-        | CosmosSdkMsg::MsgInstantiateContract {
-            init_msg: msg,
-            sender,
-            ..
-        } => msg_sender == sender && &sent_msg.to_vec() == msg,
-        CosmosSdkMsg::MsgRecvPacket {
-           packet: Packet {
-            sequence,
-            source_port,
-            source_channel,
-            destination_port,
-            destination_channel,
-            data,
-           }, ..
-        } => match handle_type {
-            HandleType::HANDLE_TYPE_IBC_PACKET_RECEIVE => {
-                let parsed_sent_msg = serde_json::from_slice::<IbcPacketReceiveMsg>(&sent_msg.msg);
-                if parsed_sent_msg.is_err() {
-                    trace!("get_verified_msg HANDLE_TYPE_IBC_PACKET_RECEIVE: sent_msg.msg cannot be parsed as IbcPacketReceiveMsg: {:?} Error: {:?}", String::from_utf8_lossy(&sent_msg.msg), parsed_sent_msg.err());
-
-                    trace!("Checking if sent_msg & data are encrypted");
-                    return &sent_msg.to_vec() == data;
-                }
-                let parsed = parsed_sent_msg.unwrap();
-
-                parsed.packet.data.as_slice() == data.as_slice()
-                    && parsed.packet.sequence == *sequence
-                    && parsed.packet.src.port_id == *source_port
-                    && parsed.packet.src.channel_id == *source_channel
-                    && parsed.packet.dest.port_id == *destination_port
-                    && parsed.packet.dest.channel_id == *destination_channel
-                    // TODO check timeout too? sequence + destination_channel + data should be enough
-            }
-            HandleType::HANDLE_TYPE_IBC_WASM_HOOKS_INCOMING_TRANSFER => {
-                let fungible_token_packet_data = serde_json::from_slice::<FungibleTokenPacketData>(data);
-                if fungible_token_packet_data.is_err(){
-                    trace!("get_verified_msg HANDLE_TYPE_IBC_WASM_HOOKS_INCOMING_TRANSFER: data cannot be parsed as FungibleTokenPacketData: {:?} Error: {:?}", String::from_utf8_lossy(data), fungible_token_packet_data.err());
-                    return false;
-                }
-                let fungible_token_packet_data= fungible_token_packet_data.unwrap();
-
-
-                let ibc_hooks_incoming_transfer_msg = serde_json::from_slice::<IbcHooksIncomingTransferMsg>(fungible_token_packet_data.memo.clone().unwrap_or_else(|| "".to_string()).as_bytes());
-                if ibc_hooks_incoming_transfer_msg.is_err(){
-                    trace!("get_verified_msg HANDLE_TYPE_IBC_WASM_HOOKS_INCOMING_TRANSFER: fungible_token_packet_data.memo cannot be parsed as IbcHooksIncomingTransferMsg: {:?} Error: {:?}", fungible_token_packet_data.memo, ibc_hooks_incoming_transfer_msg.err());
-                    return false;
-                }
-                let ibc_hooks_incoming_transfer_msg = ibc_hooks_incoming_transfer_msg.unwrap();
-                let sent_msg_value = serde_json::from_slice::<serde_json::Value>(&sent_msg.msg);
-                if sent_msg_value.is_err(){
-                    trace!("get_verified_msg HANDLE_TYPE_IBC_WASM_HOOKS_INCOMING_TRANSFER: sent_msg.msg cannot be parsed as serde_json::Value: {:?} Error: {:?}", String::from_utf8_lossy(&sent_msg.msg), sent_msg_value.err());
-                    return false;
-                }
-
-                ibc_hooks_incoming_transfer_msg.wasm.msg == sent_msg_value.unwrap()
-            }
-            _ => false,
-        },
-        CosmosSdkMsg::Other => false,
-        CosmosSdkMsg::MsgAcknowledgement { packet, acknowledgement, signer,.. } => {
-            match handle_type {
-                HandleType::HANDLE_TYPE_IBC_PACKET_ACK => {
-                    let send_msg_ack_msg = serde_json::from_slice::<IBCPacketAckMsg>(&sent_msg.msg);
-                    if send_msg_ack_msg.is_err(){
-                        trace!("get_verified_msg HANDLE_TYPE_IBC_PACKET_ACK: sent_msg.msg cannot be parsed as IBCPacketAckMsg: {:?} Error: {:?}", String::from_utf8_lossy(&sent_msg.msg), send_msg_ack_msg.err());
-                        return false;
-                    }
-                    let sent_msg_ack_msg= send_msg_ack_msg.unwrap();
-
-                    sent_msg_ack_msg.original_packet.src.channel_id == packet.source_channel &&
-                    sent_msg_ack_msg.original_packet.src.port_id == packet.source_port &&
-                    sent_msg_ack_msg.original_packet.dest.channel_id == packet.destination_channel &&
-                    sent_msg_ack_msg.original_packet.dest.port_id == packet.destination_port &&
-                    sent_msg_ack_msg.original_packet.sequence == packet.sequence &&
-                    sent_msg_ack_msg.original_packet.data.0 == packet.data &&
-                    sent_msg_ack_msg.relayer == *signer &&
-                    sent_msg_ack_msg.acknowledgement.data.0 == *acknowledgement
-                },
-                HandleType::HANDLE_TYPE_IBC_WASM_HOOKS_OUTGOING_TRANSFER_ACK => {
-                    let ibc_lifecycle_complete = serde_json::from_slice::<IBCLifecycleComplete>(&sent_msg.msg);
-                if ibc_lifecycle_complete.is_err(){
-                    trace!("get_verified_msg HANDLE_TYPE_IBC_WASM_HOOKS_OUTGOING_TRANSFER_ACK: sent_msg.msg cannot be parsed as IBCLifecycleComplete: {:?} Error: {:?}", String::from_utf8_lossy(&sent_msg.msg), ibc_lifecycle_complete.err());
-                    return false;
-                }
-                let ibc_lifecycle_complete= ibc_lifecycle_complete.unwrap();
-
-              match  ibc_lifecycle_complete {
-                IBCLifecycleComplete::IBCLifecycleComplete(IBCLifecycleCompleteOptions::IBCAck { channel, sequence, ack, success }) =>
-                    channel == packet.source_channel
-                    && sequence == packet.sequence
-                    && ack == String::from_utf8_lossy( acknowledgement)
-                    && success != is_transfer_ack_error(acknowledgement),
-                IBCLifecycleComplete::IBCLifecycleComplete(IBCLifecycleCompleteOptions::IBCTimeout { .. }) => false,
-            }
-                },
-            _ => false,
-
-            }
-        },
-        CosmosSdkMsg::MsgTimeout { packet, signer,.. } => {
-            match handle_type {
-                HandleType::HANDLE_TYPE_IBC_PACKET_TIMEOUT => {
-                    let send_msg_timeout_msg = serde_json::from_slice::<IBCPacketTimeoutMsg>(&sent_msg.msg);
-                    if send_msg_timeout_msg.is_err(){
-                        trace!("get_verified_msg HANDLE_TYPE_IBC_PACKET_TIMEOUT: sent_msg.msg cannot be parsed as IBCPacketTimeoutMsg: {:?} Error: {:?}", String::from_utf8_lossy(&sent_msg.msg), send_msg_timeout_msg.err());
-                        return false;
-                    }
-                    let sent_msg_timeout_msg= send_msg_timeout_msg.unwrap();
-
-                    sent_msg_timeout_msg.packet.src.channel_id == packet.source_channel &&
-                    sent_msg_timeout_msg.packet.src.port_id == packet.source_port &&
-                    sent_msg_timeout_msg.packet.dest.channel_id == packet.destination_channel &&
-                    sent_msg_timeout_msg.packet.dest.port_id == packet.destination_port &&
-                    sent_msg_timeout_msg.packet.sequence == packet.sequence &&
-                    sent_msg_timeout_msg.packet.data.0 == packet.data &&
-                    sent_msg_timeout_msg.relayer == *signer
-                },
-                HandleType::HANDLE_TYPE_IBC_WASM_HOOKS_OUTGOING_TRANSFER_TIMEOUT => {
-                    let ibc_lifecycle_complete = serde_json::from_slice::<IBCLifecycleComplete>(&sent_msg.msg);
-                if ibc_lifecycle_complete.is_err(){
-                    trace!("get_verified_msg HANDLE_TYPE_IBC_WASM_HOOKS_OUTGOING_TRANSFER_TIMEOUT: sent_msg.msg cannot be parsed as IBCLifecycleComplete: {:?} Error: {:?}", String::from_utf8_lossy(&sent_msg.msg), ibc_lifecycle_complete.err());
-                    return false;
-                }
-                  let ibc_lifecycle_complete= ibc_lifecycle_complete.unwrap();
-
-              match  ibc_lifecycle_complete {
-                IBCLifecycleComplete::IBCLifecycleComplete(IBCLifecycleCompleteOptions::IBCAck {..}) =>  false                ,
-                IBCLifecycleComplete::IBCLifecycleComplete(IBCLifecycleCompleteOptions::IBCTimeout { channel, sequence  }) =>
-                channel == packet.source_channel
-                && sequence == packet.sequence,
-            }
-                },
-            _ => false,
-
-            }
-        },
-    })
-}
-
-/// Check that the contract listed in the cosmwasm message matches the one in env
-fn verify_contract_address(msg: &CosmosSdkMsg, contract_address: &HumanAddr) -> bool {
-    // Contract address is relevant only to execute, since during sending an instantiate message the contract address is not yet known
-    match msg {
-        CosmosSdkMsg::MsgExecuteContract { contract, .. } => {
-            info!("Verifying contract address..");
-            let is_verified = contract_address == contract;
-            if !is_verified {
-                trace!(
-                    "Contract address sent to enclave {:?} is not the same as the signed one {:?}",
-                    contract_address,
-                    *contract
-                );
-            }
-            is_verified
-        }
-        CosmosSdkMsg::MsgInstantiateContract { .. } => true,
-        CosmosSdkMsg::Other => false,
-        CosmosSdkMsg::MsgRecvPacket {
-            packet:
-                Packet {
-                    destination_port,
-                    data,
-                    ..
-                },
-            ..
-        } => {
-            if destination_port == "transfer" {
-                // Packet was routed here through ibc-hooks
-
-                // Parse data as FungibleTokenPacketData JSON
-                let packet_data: FungibleTokenPacketData = match serde_json::from_slice(
-                    data.as_slice(),
-                ) {
-                    Ok(packet_data) => packet_data,
-                    Err(err) => {
-                        trace!(
-                            "Contract was called via ibc-hooks but packet_data cannot be parsed as FungibleTokenPacketData: {:?} Error: {:?}",
-                            String::from_utf8_lossy(data.as_slice()),
-                            err,
-                        );
-                        return false;
-                    }
-                };
-
-                // memo must be set in ibc-hooks
-                let memo = match packet_data.memo {
-                    Some(memo) => memo,
-                    None => {
-                        trace!("Contract was called via ibc-hooks but packet_data.memo is empty");
-                        return false;
-                    }
-                };
-
-                // Parse data.memo as IbcHooksWasmMsg JSON
-                let wasm_msg: IbcHooksIncomingTransferMsg = match serde_json::from_slice(
-                    memo.as_bytes(),
-                ) {
-                    Ok(wasm_msg) => wasm_msg,
-                    Err(err) => {
-                        trace!(
-                            "Contract was called via ibc-hooks but packet_data.memo cannot be parsed as IbcHooksWasmMsg: {:?} Error: {:?}",
-                            memo,
-                            err,
-                        );
-                        return false;
-                    }
-                };
-
-                // In ibc-hooks contract_address == packet_data.memo.wasm.contract == packet_data.receiver
-                let is_verified = *contract_address == packet_data.receiver
-                    && *contract_address == wasm_msg.wasm.contract;
-                if !is_verified {
-                    trace!(
-                        "Contract address sent to enclave {:?} is not the same as in ibc-hooks packet receiver={:?} memo={:?}",
-                        contract_address,
-                        packet_data.receiver,
-                        wasm_msg.wasm.contract
-                    );
-                }
-                is_verified
-            } else {
-                // Packet is for an IBC enabled contract
-                // destination_port is of the form "wasm.{contract_address}"
-
-                // Extract contract_address from destination_port
-                // This also checks that destination_port starts with "wasm."
-                let contract_address_from_port = match destination_port.strip_prefix("wasm.") {
-                    Some(contract_address) => contract_address,
-                    None => {
-                        trace!(
-                            "IBC-enabled Contract was called via MsgRecvPacket but destination_port doesn't start with \"wasm.\": {:?}",
-                            destination_port,
-                        );
-                        return false;
-                    }
-                };
-
-                let is_verified = *contract_address == HumanAddr::from(contract_address_from_port);
-                if !is_verified {
-                    trace!(
-                        "IBC-enabled Contract address sent to enclave {:?} is not the same as extracted from MsgRecvPacket but destination_port: {:?}",
-                        contract_address,
-                        contract_address_from_port,
-                    );
-                }
-                is_verified
-            }
-        }
-        CosmosSdkMsg::MsgAcknowledgement {
-            packet: Packet {
-                source_port, data, ..
-            },
-            ..
-        }
-        | CosmosSdkMsg::MsgTimeout {
-            packet: Packet {
-                source_port, data, ..
-            },
-            ..
-        } => {
-            if source_port == "transfer" {
-                // Packet was sent from a contract via the transfer port.
-                // We're getting the ack (and timeout) here because the memo field contained `{"ibc_callback": "secret1contractAddr"}`,
-                // and ibc-hooks routes the ack into `secret1contractAddr`.
-
-                // Parse data as FungibleTokenPacketData JSON
-                let packet_data: FungibleTokenPacketData = match serde_json::from_slice(
-                    data.as_slice(),
-                ) {
-                    Ok(packet_data) => packet_data,
-                    Err(err) => {
-                        trace!(
-                            "Contract was called via ibc-hooks ack callback but packet_data cannot be parsed as FungibleTokenPacketData: {:?} Error: {:?}",
-                            String::from_utf8_lossy(data.as_slice()),
-                            err,
-                        );
-                        return false;
-                    }
-                };
-
-                // memo must be set in ibc-hooks
-                let memo = match packet_data.memo {
-                    Some(memo) => memo,
-                    None => {
-                        trace!("Contract was called via ibc-hooks ack callback but packet_data.memo is empty");
-                        return false;
-                    }
-                };
-
-                // Parse data.memo as `{"ibc_callback": "secret1contractAddr"}` JSON
-                let ibc_hooks_outgoing_memo: IbcHooksOutgoingTransferMemo =
-                    match serde_json::from_slice(memo.as_bytes()) {
-                        Ok(wasm_msg) => wasm_msg,
-                        Err(err) => {
-                            trace!(
-                                "Contract was called via ibc-hooks but packet_data.memo cannot be parsed as IbcHooksWasmMsg: {:?} Error: {:?}",
-                                memo,
-                                err,
-                            );
-                            return false;
-                        }
-                    };
-
-                let is_verified = *contract_address == ibc_hooks_outgoing_memo.ibc_callback
-                    && *contract_address == packet_data.sender;
-                if !is_verified {
-                    trace!(
-                        "Contract address sent to enclave {:?} is not the same as in ibc-hooks outgoing transfer callback address packet {:?}",
-                        contract_address,
-                        ibc_hooks_outgoing_memo.ibc_callback
-                    );
-                }
-                is_verified
-            } else {
-                // Packet was sent from an IBC enabled contract
-                // source_port is of the form "wasm.{contract_address}"
-
-                // Extract contract_address from source_port
-                // This also checks that source_port starts with "wasm."
-                let contract_address_from_port = match source_port.strip_prefix("wasm.") {
-                    Some(contract_address) => contract_address,
-                    None => {
-                        trace!(
-                            "IBC-enabled Contract was called via MsgAcknowledgement but source_port doesn't start with \"wasm.\": {:?}",
-                            source_port,
-                        );
-                        return false;
-                    }
-                };
-
-                let is_verified = *contract_address == HumanAddr::from(contract_address_from_port);
-                if !is_verified {
-                    trace!(
-                        "Contract address sent to enclave {:?} is not the same as extracted from MsgAcknowledgement but source_port: {:?}",
-                        contract_address,
-                        contract_address_from_port,
-                    );
-                }
-                is_verified
-            }
-        }
-    }
-}
-
-/// Check that the funds listed in the cosmwasm message matches the ones in env
-fn verify_sent_funds(msg: &CosmosSdkMsg, sent_funds_msg: &[Coin]) -> bool {
-    match msg {
-        CosmosSdkMsg::MsgExecuteContract { sent_funds, .. }
-        | CosmosSdkMsg::MsgInstantiateContract {
-            init_funds: sent_funds,
-            ..
-        } => sent_funds_msg == sent_funds,
-        CosmosSdkMsg::Other => false,
-        CosmosSdkMsg::MsgRecvPacket {
-            packet:
-                Packet {
-                    data,
-                    source_port,
-                    source_channel,
-                    destination_port,
-                    destination_channel,
-                    ..
-                },
-            ..
-        } => {
-            if destination_port == "transfer" {
-                // Packet was routed here through ibc-hooks
-
-                // Should be just one coin
-                if sent_funds_msg.len() != 1 {
-                    trace!(
-                        "Contract was called via ibc-hooks but sent_funds_msg.len() != 1: {:?}",
-                        sent_funds_msg,
-                    );
-                    return false;
-                }
-
-                let sent_funds_msg_coin = &sent_funds_msg[0];
-
-                // Parse data as FungibleTokenPacketData JSON
-                let packet_data: FungibleTokenPacketData = match serde_json::from_slice(
-                    data.as_slice(),
-                ) {
-                    Ok(packet_data) => packet_data,
-                    Err(err) => {
-                        trace!(
-                            "Contract was called via ibc-hooks but packet_data cannot be parsed as FungibleTokenPacketData: {:?} Error: {:?}",
-                            String::from_utf8_lossy(data.as_slice()),
-                            err,
-                        );
-                        return false;
-                    }
-                };
-
-                // Check amount
-                if sent_funds_msg_coin.amount != packet_data.amount {
-                    trace!(
-                        "Contract was called via ibc-hooks but sent_funds_msg_coin.amount != packet_data.amount: {:?} != {:?}",
-                        sent_funds_msg_coin.amount,
-                        packet_data.amount,
-                    );
-                    return false;
-                }
-
-                // The packet's denom is the denom in the sender chain.
-                // It needs to be converted to the local denom.
-                // Logic source: https://github.com/scrtlabs/SecretNetwork/blob/96b0ba7d6/x/ibc-hooks/wasm_hook.go#L483-L513
-                let denom: String = if receiver_chain_is_source(
-                    source_port,
-                    source_channel,
-                    &packet_data.denom,
-                ) {
-                    // remove prefix added by sender chain
-                    let voucher_prefix = get_denom_prefix(source_port, source_channel);
-
-                    let unprefixed_denom: String = match packet_data
-                        .denom
-                        .strip_prefix(&voucher_prefix)
-                    {
-                        Some(unprefixed_denom) => unprefixed_denom.to_string(),
-                        None => {
-                            trace!(
-                                "Contract was called via ibc-hooks but packet_data.denom doesn't start with voucher_prefix: {:?} != {:?}",
-                                packet_data.denom,
-                                voucher_prefix,
-                            );
-                            return false;
-                        }
-                    };
-
-                    // The denomination used to send the coins is either the native denom or the hash of the path
-                    // if the denomination is not native.
-                    let denom_trace = parse_denom_trace(&unprefixed_denom);
-                    if !denom_trace.path.is_empty() {
-                        denom_trace.ibc_denom()
-                    } else {
-                        unprefixed_denom
-                    }
-                } else {
-                    let prefixed_denom = get_denom_prefix(destination_port, destination_channel)
-                        + &packet_data.denom;
-                    parse_denom_trace(&prefixed_denom).ibc_denom()
-                };
-
-                // Check denom
-                if sent_funds_msg_coin.denom.to_lowercase() != denom.to_lowercase() {
-                    trace!(
-                        "Contract was called via ibc-hooks but sent_funds_msg_coin.denom != denom: {:?} != {:?}",
-                        sent_funds_msg_coin.denom,
-                        denom,
-                    );
-                    return false;
-                }
-
-                true
-            } else {
-                // Packet is for an IBC enabled contract
-                // No funds should be sent
-                sent_funds_msg.is_empty()
-            }
-        }
-        CosmosSdkMsg::MsgAcknowledgement { .. } => {
-            // No funds should be sent with a MsgAcknowledgement
-            sent_funds_msg.is_empty()
-        }
-        CosmosSdkMsg::MsgTimeout { .. } => {
-            // No funds should be sent with a MsgTimeout
-            sent_funds_msg.is_empty()
-        }
-    }
-}
-
 fn verify_message_params(
     messages: &[CosmosSdkMsg],
     sender: &CanonicalAddr,
@@ -1145,24 +662,8 @@ fn verify_message_params(
     let msg = msg.unwrap();
 
     info!("Verifying message sender...");
-    match msg {
-        CosmosSdkMsg::MsgRecvPacket { .. }
-        | CosmosSdkMsg::MsgAcknowledgement { .. }
-        | CosmosSdkMsg::MsgTimeout { .. } => {
-            // No sender to verify.
-            // Going to pass null sender to the contract if all other checks pass.
-        }
-        CosmosSdkMsg::MsgExecuteContract { .. }
-        | CosmosSdkMsg::MsgInstantiateContract { .. }
-        | CosmosSdkMsg::Other => {
-            if msg.sender() != Some(sender) {
-                warn!(
-                    "message sender did not match cosmwasm message sender: {:?} {:?}",
-                    sender, msg
-                );
-                return false;
-            }
-        }
+    if let Some(value) = verify_sender(msg, sender) {
+        return value;
     }
 
     info!("Verifying contract address...");
