+++ cosmwasm/enclaves/shared/contract-engine/src/lib.rs
@@ -17,6 +17,7 @@ pub mod external;
 mod gas;
 mod ibc_denom_utils;
 mod ibc_message;
+mod input_validation;
 mod io;
 mod message;
 mod message_utils;
